+++ packages/common/src/index.ts
@@ -7,6 +7,7 @@ export * from './utils/test-utils'
 export * from './logger-provider'
 export * from './logger-provider-old' // TODO remove this
 export * from './loggers'
+export * from './networks'
 export * from './pinning'
 export * from './unreachable-case-error'
 export * from './running-state-like'