+++ packages/logger/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](/compare/@ceramicnetwork/logger@0.1.0...@ceramicnetwork/logger@0.1.1) (2021-04-02)
+
+
+### Bug Fixes
+
+* **common, logger:** Clean up dependencies (#1164) 191ad31, closes #1164
+
+
+
+
+
 ## [0.1.1-rc.3](/compare/@ceramicnetwork/logger@0.1.0...@ceramicnetwork/logger@0.1.1-rc.3) (2021-03-26)